Refactor du moteur physique + uniformisation ui éditeur
Finition descriptions actions/comps
Refactor Atlas édit + right panel game objects pour utiliser le settingsButton
Refactoring du moteur physique, -> nombres décimaux à la place de nombre entier -> augmentation de la précision et compatibilité avec le delta time -> support de n'importe quel nombre de FPS.
Corrections bugs
...

+++ core/src/main/kotlin/be.catvert.pc/actions/TeleportAction.kt
@@ -21,7 +21,7 @@ class TeleportAction(var teleportPoint: Point) : Action(), CustomEditorImpl {
     }
 
     override fun insertImgui(label: String, gameObject: GameObject, level: Level, editorSceneUI: EditorScene.EditorSceneUI) {
-        ImguiHelper.point(::teleportPoint, Point(), Point(level.matrixRect.width, level.matrixRect.height), editorSceneUI)
+        ImguiHelper.point(::teleportPoint, Point(), Point(level.matrixRect.width.toFloat(), level.matrixRect.height.toFloat()), editorSceneUI)
     }
 
     override fun toString() = super.toString() + " - $teleportPoint"
