+++ core/src/main/kotlin/be.catvert.pc/utility/Point.kt
@@ -4,7 +4,7 @@ package be.catvert.pc.utility
 /**
  * Classe de donnée représentant un point dans l'espace
  */
-data class Point(val x: Int = 0, val y: Int = 0) {
+data class Point(val x: Float = 0f, val y: Float = 0f) {
     constructor(point: Point) : this(point.x, point.y)
 
     override fun toString(): String = "{ $x ; $y }"
